// packages/components/src/ui/Badge/Badge.styles.d.ts
export declare const useStyles: () => {
  badge: {
    borderRadius: import('packages/theme/src/constants/Radius').Radius;
    marginRight: string;
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    justifyContent: 'center';
  };
  badgeInfo: {
    color: '#0264C7';
    backgroundColor: '#E5F1FE';
  };
  badgeSuccess: {
    color: '#1B7440';
    backgroundColor: '#B9FDD5';
  };
  badgeError: {
    color: '#C62416';
    backgroundColor: '#FDEAE8';
  };
  badgeWarning: {
    color: '#BC4401';
    backgroundColor: '#FFEEE5';
  };
  badgeNew: {
    color: '#7F4214';
    backgroundColor: '#FBF2C6';
  };
  badgeDefault: {
    color: '#373A3F';
    backgroundColor: '#EFF1F6';
  };
  badgeSm: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
    textTransform: 'uppercase';
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    gap: import('@alveole/theme').Spacing;
  };
  badgeMd: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
    textTransform: 'uppercase';
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    gap: import('@alveole/theme').Spacing;
  };
};
//# sourceMappingURL=Badge.styles.d.ts.map